--- plugins/reference-assets/core/src/assets/choice/choices-entry/index.ts
@@ -1,1 +1,0 @@
-export * from './types';
